requirements.txt
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-jazzmin==3.0.0
django-silk==5.1.0
djangorestframework==3.15.1
drf-yasg==1.21.7
factory-boy==3.3.0
gunicorn==22.0.0
Pillow==10.3.0
psycopg2-binary==2.9.9
pytest-django==4.8.0
python-dotenv==1.0.1
pytz==2024.1